requirements.txt
# WRDS
wrds

# PostgreSQL
sqlalchemy
psycopg2-binary # Driver for Postgres

# Data Transformation & Analysis
pandas
polars           #
pyarrow          # Required for polars to_pandas conversion
statsmodels      #

# Utils
python-dotenv    # For loading .env files

# Orchestration & Testing
pytest           #